Move some helper classes into jars/fiji-lib.jar
SIOX had a problem with accessing the java.awt.geom.Shape instance
in a ShapeRoi, as the relevant method is package-local in ImageJ,
and you can only access such from a different class when it is loaded
via the _same_ class loader.

So we move the stuff into a library which is not loaded via the
plugin classloader: jars/fiji-lib.jar.

While at it, make this library a starting point for common functions
to be used by all kinds of Fiji plugins.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>