Abstract tool: make unregister check safer
Unfortunately, there is no way to detect whether the custom tools have
all been cleared (by installing a new macro set, for example).

Therefore we need to use the ugly trick to compare whether the string
_reference_ to the tool name is the same as when registering. As the
Toolbar class only has a public method to enquire about the name of the
current tool, not the name of a tool for a given id, we might hold on to
the listeners for quite some time unnecessarily.

The string cannot be the same instance as returned by getToolName()
because we always have an icon that we add, and Toolbar.addTool()
removes it again before storing the name.

We might be able to change this at a later date, if ImageJ grows the
required public API.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>