Abstract Tool: major design change
Rather than making every tool a mouse, mouse motion and mouse wheel
listener, make it none of these by default. If a subclass implements
any of these listener interfaces, a proxy is installed that calls
the respective listener methods when appropriate.

In cases where implementing the interfaces in the subclass is
inconvenient (e.g. when only one method needs to be implemented and,
say, an inner class extending the MouseAdapter class is preferable),
the listener can still be registered, e.g. by an initializer block
like this:

	{
		mouseProxy = new MouseProxy(new MyListener());
	}

This design change makes things easier for users of the base class,
as existing listeners can be reused, and no new interfaces need to be
learnt.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>