Let instances of AbstractTool be KeyListeners, too
When a subclass of AbstractTool implements the KeyListener interface,
we assume that the tool wants to get keyboard events from the image
window and canvas.

However, there is a caveat: ImageJ's default keyboard handler does not
care whether we handled a keyboard event already or not, and with the
shortcut creep in ImageJ, it is increasingly difficult to find an unused
key (and if such an unused key is found, there is no guarantee that it
will be unused in the next ImageJ version).

Therefore, we graft a proxy on top of ImageJ's default keyboard handler
which forwards only unconsumed keyboard events. Of course this means
that subclasses of the AbstractTool class _must_ call the KeyEvent's
consume() method when the event was handled!

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>