AbstractTool: a MouseMotionListener overrides ImageJ's default one
When a tool implements the MouseMotionListener interface, ImageJ should
not react to mouse motion when that tool is active.

Therefore, use the same trick (ProxyIfNotConsumed) as with the
KeyListeners to switch off ImageCanvas' MouseMotionListener if the
event has been consumed already.

For example, one would implement a tool changing the cursor thusly:

	public class Example extends AbstractTool
			implements MouseMotionListener {
		public void mouseMoved(MouseEvent e) {
			if (getOffscreenX(e) < 10)
				getImageCanvas(e).setCursor(hand);
			else
				getImageCanvas(e).setCursor(cross);
			e.consume();
		}

		public void mouseDragged(MouseEvent e) {
			getImageCanvas(e).setCursor(move);
			e.consume();
		}
	}

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>