Add a JNI helper class
An earlier attempt tried to put this into jars/Fiji.jar (which is loaded
by the system class loader) to be able to load a library exactly once
(even if the FijiClassLoader is thrown away and a new one is used, and
with it any contained singleton keeping track of the libraries that
were already).

However, that does not work. In their infinite wisdom, the Java
designers decided that a native library can only be used from within the
same classloader as the class that called System.loadLibrary(). (If the
classloaders are different, you get a rather misleading
UnsatisfiedLinkError.)

Unfortunately, you cannot unload native libraries, and neither reload
them, so this helper is only half as useful as it was meant to be: if
you call Help>Refresh Menus, the class will simply fail to load
libraries, not warn you that they have been loaded already and
unfortunately cannot be used any longer.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>