fiji.Debug.run(): handle Maven projects gently
In Maven projects, we want to add the target/ directory as plugins.dir
so that the .jar file can be found by ImageJ 1.x and parsed for a
plugins.config. However, there is also target/classes/ which ImageJ 1.x
parses, therefore it detects a duplicate plugin -- erroneously. Rather
than error out, let's just redirect the error message to the log and
move on.

We will have to fix this properly at a later stage, in ij-legacy
(because ImageJ 1.x is unlikely to fix this properly).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>